separate avatar and useravatar tests
