+++ webui/react/src/components/Table.tsx
@@ -1,7 +1,7 @@
 import { Space, Tooltip } from 'antd';
 import React from 'react';
 
-import Avatar from 'components/Avatar';
+import UserAvatar from 'components/UserAvatar';
 import Badge, { BadgeType } from 'components/Badge';
 import { ConditionalWrapper } from 'components/ConditionalWrapper';
 import HumanReadableNumber from 'components/HumanReadableNumber';
@@ -92,7 +92,7 @@ export const tooltipRenderer: Renderer = text => (
 );
 
 export const userRenderer: Renderer<{ userId: number }> = (_, record) => (
-  <Avatar userId={record.userId} />
+  <UserAvatar userId={record.userId} />
 );
 
 /* Command Task Table Column Renderers */
