+++ webui/react/src/pages/ModelDetails/ModelHeader.tsx
@@ -2,7 +2,7 @@ import { LeftOutlined } from '@ant-design/icons';
 import { Alert, Breadcrumb, Button, Dropdown, Menu, Modal, Space } from 'antd';
 import React, { useCallback, useMemo } from 'react';
 
-import Avatar from 'components/Avatar';
+import Avatar from 'components/UserAvatar';
 import InfoBox, { InfoRow } from 'components/InfoBox';
 import InlineEditor from 'components/InlineEditor';
 import Link from 'components/Link';
