+++ webui/react/src/shared/components/Avatar/index.ts
@@ -1,0 +1,3 @@
+export * from './Avatar';
+import Avatar from './Avatar';
+export default Avatar;
