+++ webui/react/src/shared/components/AvatarCard/index.ts
@@ -1,0 +1,3 @@
+export * from './AvatarCard';
+import AvatarCard from './AvatarCard';
+export default AvatarCard;
